+++ examples/commit/indicators_commit.py
@@ -9,6 +9,7 @@ import sys
 
 """ custom """
 from threatconnect import ThreatConnect
+from threatconnect.Config.IndicatorType import IndicatorType
 from threatconnect.Config.ResourceType import ResourceType
 
 # configuration file
@@ -76,6 +77,9 @@ def main():
 
     # (Optional) iterate through all results if retrieve was used above
     for res in resources:
+
+        if res.indicator is None:
+            continue
 
         # (Optional) match a particular resource by ID, Name or any other supported attribute.
         if lu_indicator == res.indicator or lu_indicator in res.indicator:
@@ -175,7 +179,9 @@ def main():
     #
 
     # this requires that the resource was instantiated at the beginning of the script.
-    resource = resources.add('4.3.254.{0:d}'.format(rn), owner)
+    s = '4.3.1.{0:d}'.format(rn)
+    print "S IS {}".format(s)
+    resource = resources.add(s, owner=owner, type=IndicatorType.ADDRESSES)
     resource.set_confidence(rn)
     resource.set_rating(randint(1, 5))
 
@@ -262,7 +268,7 @@ def main():
     #
 
     # this requires that the resource was instantiated at the beginning of the script.
-    resource = resources.add('{0!s}_{1!s}.com'.format(prefixes['host'], str(rn).zfill(3)), owner)
+    resource = resources.add('{0!s}{1!s}.com'.format(prefixes['host'], str(rn).zfill(3)), owner)
 
     resource.set_confidence(rn)
     resource.set_rating(randint(1, 5))